enh: move ``init_bold_preproc_report_wf`` into outputs module.
